[PiranhaJava] Refactor Piranha's test suite into multiple files.
We divide the massive `XPFlagCleanerTest` junit test suite into
four separate test suites, each dealing with a different testing
concern:

* CorePiranhaTest for core functionality and running the various
  test files inside test/resources/...
* ConfigurationTest for all tests related to Piranha's EP parameters
  and parsing of the `properties.json` configuration file.
* TestCaseCleanUpTest for Piranha logic related to cleaning up unit
  tests in refactored code which depend on or reference stale flags
* TreatmentGroupsTest for Piranha logic handling treatment group
  information.

In the future, if we support, say, experiment parameters, we should
add a new test suite for it.

Not only having separate files rather than a single 1 KLoc file
(with 60+ tests) makes it easier to read through and find existing
unit tests, but this should also speed our build, as gradle can
run different junit test classes in parallel.
